Fixing rbenv init
Not sure how this was working before, but ~/.bashrc has an explicit check for non-interactive `source` calls, so it's never actually running the rbenv init. We're just doing it explicitly in this script... which makes it seem more and more like this script belongs directly in the `blog.nilenso.com` repo. :/