Add new design for people page
We use CSS grids to make this work.

We introduce a `shortbio` field in the people data file that will
contain the short versions of everyone's bios. A Tera macro will iterate
over this data in chunks to fill up the grid tiles. In between these
chunks is the misc content, like tweets, images, whatever.

This tile data currently sits in `people.toml`, but can be moved
elsewhere rather easily.

Helped-by: Vikram Chintalapati <vikram@nilenso.com>